backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdf2docx==0.5.8
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
python-docx==1.1.2
